+++ rbtree.h
@@ -113,7 +113,7 @@ struct rb_root
 
 #define RB_ROOT (struct rb_root){ (struct rb_node *)0, }
 #define	rb_entry(ptr, type, member) \
-	((type *)((char *)(ptr)-(unsigned long)(&((type *)0)->member)))
+	((type *)((char *)(ptr)-(size_t)(&((type *)0)->member)))
 
 #ifdef __cplusplus
 extern "C"
